On a way of restructuring for <path d>'s 'a' instruction.
