+++ .gitignore
@@ -49,4 +49,6 @@ build-iPhoneSimulator/
 # unless supporting rvm < 1.11.0 or doing something fancy, ignore this:
 .rvmrc
 
+# project specific
 .*.swp
+*.svg
